Reverting Kevin's -l change since it wasn't completed. If we want to change that, we also need to change:
   while ((tmpint = getopt_long(argc, argv, "bc::C:f::F:hl:rs::t:v",

to not require an argument on l, as well as:

   {"learn-ir", required_argument, 0, 'l'},

to not require an argument on --learn-ir.

And finally, we need to be consistent in requiring this or not requiring this information is desired from a file. Since all other such options (-C, -F, etc.) act this way, we should keep this consistent unless we want to change all of them. Judging by our current (small) userbase, that doesn't seem like the right option, but either way, it'll need more discussion, and then based on that choice, it needs to be done completely.

- Phil