This gets firmware working from the website... here was the problem...
The first 6 bytes of the firmware file we receive from the
website will be blanked out (0xFF), and we need to fill them
in with the magic 6 bytes from the existing firmware.

So why don't we always do this? If the user has a dump from us,
it already has the right first 6 bytes... and if somehow the
firmware on the device is messed up, we don't want to ignore
that useful data in the file.

So we only retreive it from the remote if it isn't given to us.
For most users, that will be all the time.


Of course - we still don't FULLY support firmware updates because we don't tell the website when we're done. I'll get to that soon.


libconcord.cpp:
  - Make the magic 6 bytes right

- Phil