From: Scott Talbert <swt@techie.net>
Certain remotes (e.g., the 785) do not get a successful return from a reset
command even though the reset succeeds.  This patch causes this specific error
to be ignored and the reset function returns successfully.  This fixes Debian
bug #584896.

[Added by Phil, copied from email from Scott, on the usb_strerror/strerror part
of the patch:]

Well, it turns out that usb_strerror() appears to have a bug of some sort where
it is not providing the correct error message.  It is a providing a message
"could not detach kernel driver from interface 0: No data available" which
doesn't make any sense.  Using strerror() seems to provide the correct error
message which makes sense "no such device" as the device has already gone away.

Signed-off-by: Scott Talbert <swt@techie.net>
Signed-off-by: Phil Dibowitz <phil@ipom.com>