Add support for Harmony Link.
The Harmony Link is an MH "remote."  Essentially it is a WiFi to IR bridge
where you can issue commands on your smartphone via WiFI and the Link will
convert them to IR to control your devices.  In order to support the Link,
we had to add a few API calls to allow configuration of WiFI (SSID, key, etc)
as well as account configuration (email, etc).  The Link does not receive its
configuration via USB like other remotes - it receives it via WiFi.

Signed-off-by: Scott Talbert <swt@techie.net>
Signed-off-by: Phil Dibowitz <phil@ipom.com>