Fix startup of USBNet DHCP server by udev
A change was made to udev where it now kills off any processes that are in the
same cgroup that have been running for more than a few seconds (thus our DHCP
server was getting killed off).  Instead ask 'at' to launch our DHCP server so
that udev does not kill it off.

Signed-off-by: Scott Talbert <swt@techie.net>
Signed-off-by: Phil Dibowitz <phil@ipom.com>