Support the Harmony Touch (v13)
Add the PID for the Touch to the list of MH PIDs.
Add the Touch's Skin/Arch to the remote_info structs.
Add support for handling the Touch's non-standard serial number, and provide
a libconcord call to expose this information.
Don't try to read the /cfg/usercfg on the Link/Touch.
Remove set_configuration - doesn't seem to be needed and causes problems for
Harmony Touch.
Add new API methods so that MHGUI can perform a sync operation.
Fix WriteFile() so that it reads and responds to the periodic acks.
Implement SetTime() for the Touch.

Signed-off-by: Scott Talbert <swt@techie.net>
Signed-off-by: Phil Dibowitz <phil@ipom.com>